bug 15266 fixed : commandline delete item, will get error.
git-svn-id: http://talendforge.org/svn/top/trunk@47640 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.dataprofiler.core/src/org/talend/dataprofiler/core/ui/action/actions/DeleteFolderAction.java
@@ -65,7 +65,17 @@ public class DeleteFolderAction extends Action {
             return;
         }
         try {
-            folder.delete(true, null);
+            if (isDeleteForever) {
+                // MOD qiongli 2010-8-5,bug 14697.
+                delsubFolderForever(folder);
+                if (LogicalDeleteFileHandle.isStartWithDelFolder(folder.getFullPath().toOSString())) {
+                    LogicalDeleteFileHandle.replaceInFile(LogicalDeleteFileHandle.folderType + folder.getFullPath().toOSString(),
+                            PluginConstant.EMPTY_STRING);
+                    folder.delete(true, null);
+                }
+            } else {
+                LogicalDeleteFileHandle.saveElement(LogicalDeleteFileHandle.folderType, folder.getFullPath().toOSString());
+            }
             CorePlugin.getDefault().refreshDQView();
             ProxyRepositoryManager.getInstance().save();
         } catch (CoreException e) {
